% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{longest_distance}
\alias{longest_distance}
\title{Longest distance}
\usage{
longest_distance(data, shipname)
}
\arguments{
\item{data}{dataframe containing at least the following columns : SHIPNAME, LAT, LON, DATETIME}

\item{shipname}{a string}
}
\description{
Return the longest distance for a given ship
}
\examples{
longest_distance(ships, "KAROLI")
}
