% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ships}
\alias{ships}
\title{AIS Dataset}
\format{
A data frame with 3102887 rows and 20 variables:
\describe{
  \item{LAT}{ship’s latitude}
  \item{LON}{ship’s longitude}
  ...
}
}
\usage{
ships
}
\description{
A dataset containing marine data transmitted by the AIS
(Automatic identification system) standard.
}
\keyword{datasets}
